+++ src/per_test_config.rs
@@ -1,5 +1,6 @@
-//! This module allows you to configure the default settings for all
-//! tests. All data structures here are normally parsed from `@` comments
+//! This module allows you to configure the default settings for all tests.
+//!
+//! All data structures here are normally parsed from `@` comments
 //! in the files. These comments still overwrite the defaults, although
 //! some boolean settings have no way to disable them.
